Attach Bluetooth device to VM
The BT device needs to be attached to the VM. We use BT for various
things in Home Assistant so it's necessary for it to be available to the
VM.

This makes the assumption that there will be only one USB device
detected and that it will be for the BT device. So far on two separate
computers this has been true so we'll stick with this for now.