Use more robust method of attaching USB device
Using `usbattach` required the VM to be started which we don't do as
part of this script.

Rather than attaching the USB device to the VM directly, we're using a
USB filter to do so. We're also selecting the first USB device so that
even if `vboxmanage list usbhost` returns more than one USB device,
we'll only use the first. This isn't guaranteed to be the Bluetooth
device, but it will at least make the resulting variable always valid as
it couldn't possibly be multiple IDs.