Don't automatically add usb filter to VM
The method we're using to attach the USB device to the VM is too naive.
It assumes the first device will be the one we want (a BT device). This
is too unreliable so for more we'll just require the filter being added
manually or the USB device attached manually.